+++ include/linux/hisi/usb/hisi_usb_interface.h
@@ -15,6 +15,6 @@ int dwc3_misc_ctrl_get(enum misc_ctrl_type type);
 void dwc3_misc_ctrl_put(enum misc_ctrl_type type);
 volatile unsigned int hisi_dwc3_usbcore_read(u32 offset);
 void hisi_dwc3_usbcore_write(u32 offset, u32 value);
-u16 usb31phy_cr_read(u32 addr);
-int usb31phy_cr_write(u32 addr, u16 value);
+u32 usb31phy_cr_read(u32 addr);
+int usb31phy_cr_write(u32 addr, u32 value);
 #endif /* hisi_usb_interface.h */
